Refactor futures module into futures packages
