+++ src/pydvl/utils/parallel/futures/ray.py
@@ -5,46 +5,19 @@ import threading
 import time
 import types
 from concurrent.futures import BrokenExecutor, Executor, Future
-from contextlib import contextmanager
 from dataclasses import asdict
-from typing import Any, Callable, Generator, Optional, TypeVar
+from typing import Any, Callable, Optional, TypeVar
 from weakref import ref
 
 import ray
 
-from ..config import ParallelConfig
+from pydvl.utils import ParallelConfig
 
-__all__ = ["init_executor", "RayExecutor"]
+__all__ = ["RayExecutor"]
 
 T = TypeVar("T")
 
 logger = logging.getLogger(__name__)
-
-
-@contextmanager
-def init_executor(
-    config: ParallelConfig = ParallelConfig(),
-) -> Generator[Executor, None, None]:
-    """Initializes a futures executor based on the passed parallel configuration object.
-
-    :param max_workers: Maximum number of concurrent tasks.
-    :param config: instance of :class:`~pydvl.utils.config.ParallelConfig` with cluster address, number of cpus, etc.
-
-    :Example:
-
-    >>> from pydvl.utils.parallel.futures import init_executor
-    >>> from pydvl.utils.config import ParallelConfig
-    >>> config = ParallelConfig(backend="ray")
-    >>> with init_executor(config=config) as executor:
-    ...     pass
-
-    """
-    if config.backend == "ray":
-        max_workers = config.n_workers
-        with RayExecutor(max_workers, config=config) as executor:
-            yield executor
-    else:
-        raise NotImplementedError(f"Unexpected parallel type {config.backend}")
 
 
 class RayExecutor(Executor):
